backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
langchain
python-multipart
cohere
faiss-cpu
python-dotenv
google-generativeai
